hooks/pkglint-devel-paths: check even if no -devel subpkg exists yet
Typically when adding a new package it's easier to build it once as a
single package and then start adding -devel subpackage based on files in
the first created package.

It simply makes sense to get these warnings so a -devel subpackage can
be added easily without annoying build errors and guessing whether or
not you need one.

NOTE: This will produce some false-positives such as with header-only
library packages!